Simplify TorchCodec benchmark video-dir API
